refactor(tangle-dapp): Minor changes
